+++ lib/duffer.dart
@@ -3,7 +3,6 @@ library duffer;
 
 export 'src/bytebuf_base.dart';
 export 'src/bytebuf_allocator.dart';
-export 'src/bytebuf_iterator.dart';
 export 'src/extensions.dart';
 export 'src/pooled.dart';
 export 'src/unpooled.dart';